refactor: move active surface playback scenarios
